style: change the dashboards to accordion menu
