+++ apps/web/src/app/[lang]/(main)/company/layout.tsx
@@ -55,7 +55,7 @@ export default function Layout({ children }: LayoutProps) {
         defaultActiveSectionId={path}
         openOnNewPage={true}
         content={children}
-        contentClassName="flex flex-col-reverse md:flex-row flex-wrap-reverse flex-1 lg:gap-16 md:gap-4 justify-center w-[80vw]"
+        contentClassName="flex flex-col-reverse md:flex-row flex-wrap-reverse flex-1 lg:gap-16 md:gap-4 justify-center"
       />
     </div>
   );
