--- apps/web/src/app/[lang]/(main)/dashboard/layout.tsx
@@ -1,76 +1,0 @@
-"use client";
-
-import { SectionLayout } from "@repo/ayasofyazilim-ui/templates/section-layout";
-import { usePathname, useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-const navbarItems = [
-  {
-    name: "Roles",
-    id: "role",
-  },
-
-  {
-    name: "Users",
-    id: "user",
-  },
-  {
-    name: "Editions",
-    id: "edition",
-  },
-
-  {
-    name: "Tenants",
-    id: "tenant",
-  },
-
-  {
-    name: "Claim Type",
-    id: "claimType",
-  },
-
-  {
-    name: "Applications",
-    id: "applications",
-  },
-
-  {
-    name: "Scopes",
-    id: "scopes",
-  },
-
-  {
-    name: "Languages",
-    id: "languages",
-  },
-];
-function isPathValid(path: string) {
-  const validPaths = navbarItems.map((item) => item.id);
-  return validPaths.includes(path) || path === undefined;
-}
-
-type LayoutProps = {
-  children: JSX.Element;
-};
-
-export default function Layout({ children }: LayoutProps) {
-  const router = useRouter();
-  const pathname = usePathname();
-  const path = pathname.split("dashboard/")?.[1];
-
-  useEffect(() => {
-    if (!path || isPathValid(path) === false) {
-      router.push("dashboard/role");
-    }
-  }, []);
-
-  return (
-    <SectionLayout
-      sections={navbarItems}
-      defaultActiveSectionId={path}
-      openOnNewPage={true}
-      content={children}
-      contentClassName="flex flex-col-reverse md:flex-row flex-wrap-reverse flex-1 lg:gap-16 md:gap-4 justify-center w-[80vw]"
-    />
-  );
-}
