--- apps/web/src/app/[lang]/(main)/dashboard/page.tsx
@@ -1,6 +1,0 @@
-import { redirect } from "next/navigation";
-import { getBaseLink } from "src/utils";
-
-export default function Page() {
-  redirect(getBaseLink("/dashboard/role"));
-}
